[BE] feat: 댓글 관련 기능 (#60)
* chore: 댓글 추가 시 response 제거(#58)

* feat: 댓글 추가 기능(#58)

* feat: 댓글 삭제 기능(#58)

댓글 soft delete, 해당 댓글에 달린 이모지 데이터 삭제

* feat: @Modifying과 @Query, nativeQuery로 N번 발생하는 쿼리 메서드 리팩토링하기(#58)

* feat: 댓글 수정 기능 추가(#58)

* feat: 변하지 않는 데이터를 @Component, @PostConstruct 어노테이션을 사용해 빈으로 등록(#58)

* feat: response 형식에 맞도록 데이터 자료구조 추가(#58)

* feat: 변하지 않는 데이터들은 서버 실행시 한 번만 조회하도록 함(#58)

* feat: 댓글 목록 조회 기능을 위해 쿼리 구현체 생성(#58)

* feat: 댓글 목록 조회 기능 구현(#58)

추후 서브 쿼리를 사용하여 리팩토링

* chore: 가독성을 위해 포맷 수정(#58)

* feat: 이모지 선택 기능 추가 (#58)

+++ src/main/java/reviewme/be/util/repository/EmojiRepository.java
@@ -5,7 +5,7 @@ import reviewme.be.util.entity.Emoji;
 
 import java.util.List;
 
-public interface EmojiRepository extends JpaRepository<Emoji, Long> {
+public interface EmojiRepository extends JpaRepository<Emoji, Integer> {
 
     List<Emoji> findAll();
 }
